refactor: move mock code into new mock mod
This moves all "extra" functionality into a new
module "mock" with a corresponding feature flag
and its own Error enum.

The moved items are:
 GenesisBuilderMock --> mock::GenesisBuilder
 SpentbookNodeMock  --> mock::SpentbookNode
 SimpleKeyManager   --> mock::KeyManager
 SimpleSigner       --> mock::Signer

This makes the core library leaner, meaner and easier to audit.

Cargo changes:
* add 'mock' feature flag
* set mock feature flag when building tests

Code changes:
* use mock::* in tests, reissue, mint-repl
* move GenesisBuilderMock into mock::GenesisBuilder
* remove crate::Error::Spentbook*
* add crate::Error::Mock, only when mock feature enabled.
* move SimpleKeyManager, SimpleSigner into mock::{KeyManager, Signer}
* expose pub crate::mock in lib.rs when feature enabled
* move SpentbookNodeMock into mock::SpentbookNode
* add mock::Error
