+++ src/mock/error.rs
@@ -1,0 +1,22 @@
+// Copyright 2022 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+use thiserror::Error;
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Error, Debug, Clone)]
+/// Mock error variants.
+pub enum Error {
+    #[error("Key image has already been spent")]
+    KeyImageAlreadySpent,
+
+    #[error("The transaction input has {0:?} public keys but found {1:?} matching outputs in spentbook.")]
+    RingSizeMismatch(usize, usize),
+}
