+++ src/mock/mod.rs
@@ -1,0 +1,19 @@
+// Copyright 2022 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+mod error;
+mod genesis_builder;
+mod key_manager;
+mod spentbook;
+
+pub use self::{
+    error::Error,
+    genesis_builder::GenesisBuilder,
+    key_manager::{KeyManager, Signer},
+    spentbook::SpentBookNode,
+};
